Add degree/radian mode, improve button layout
Add a button for switching between degree and radian mode.
Handle display updates in response to mode switching in a reasonoble way.
Improve the button layout so that it is usable on tablets.
Add DEG/RAD mode display at the top.
Fiddle with some of the layout parameters so they make more sense
to me.

This may need further tuning.

Change-Id: I8d04b235cfcb1f0a0d8f4a0ad6a0dc6e5eac8c5d