Fix result overlap, clear logic, display bugs, etc.
Adjust result padding to prevent overlapping with options menu,
resulting in inability to access options menu.

Reset state correctly after clear, preventing occasional crashes.
Do this at end of animation to prevent animation glitches.
Have clear but the display in a state in which COPY does
nothing interesting.

Treat a not-known-to-be-exact zero result better.  Before this
we displayed it as just "0", giving the mistaken impression of
exactness.  It is unclear whether we should display it as "0.00000000"
or add an exponent to indicate the real evaluation precision.
The former probably works better for beginners; the latter for
experts.  We now do the former, which is probably right.

Also correctly save and restore not-yet analyzed characters.

Change-Id: Iac46d0d3b5a06f61beb6af7d4f0910e43284f52b