Improve copy/paste handling, minor instant result fixes
Bug: 20764417
Bug: 20838790

Finish the action mode when we do almost anything else.
We should probably be even more aggressive about terminating it,
and we may eventually want to use a different approach.
But this avoids the crashes, and greatly improves the behavior.

Display copy menu only when there is something to copy.

Fix a couple of bugs introduced by recent instant display overhaul
(one line each):

- Instant display disappeared on rotation.

- It was sometimes still possible to scroll the previous result after
CLR.

Change-Id: I1a91d312358898add1e281aaba116d709b4c92a7