Prevent scrolling of finite results.
Bug: 20562484

Inhibit scrolling past the last nonzero digit when we can identify it.
Increase length of immediate result a bit, to make it closer to the
L calculator.  On "=", expand less for scrollable results, more for
short non-scrollable results.  This allows us to use larger fonts for
short finite results.

Fix up animation logic when displaying result.  The old version was
never quite right, and became more visibly wrong with variable
enlargement.  This version is simpler and seems to give much better
results.

This still does not add ellipsis at the right end of a result.  But
it is now easily possible to tell whether a result is known finite by
attempting to scroll it.  That may be good enough.

Remove some obsolete TODO entries.

Change-Id: I25a842a743e1c27191ca18ac69aa9eef0f0ea9b1