Use correct decimal separator in result.
Bug: 21282652

Consistently use key labels to internationalize result.

This makes the Farsi locale, and possibly others where we diverge
from the normal digit representation specified by the locale, work
sanely.

It remains unclear whether this divergence is a bug or a feature.

Don't statically initialize mDecimalPt, since it is already initialized
when needed, and may need to be reinitialized with locale changes.

Always recognize ',' as a decimal separator.

Remove mFactChar, since it doesn't seem to buy us anything.

Change-Id: I3f7c413d380e009fc562903dc23475a7d7cd6d0a