Force textAlignment == viewEnd
Bug: 21553173

Since the Calculator is UI is predominately LTR, the error strings
should be forced to right align regardless if they contain RTL
characters. However in order to use the textAlignment attribute, the
supportsRtl attribute must also be set to "true" in the manifest and
the Calulator's layoutDirection must be hardcoded to "ltr".

Change-Id: I40396622c334e276bda9e0020dcdcfe94192a42c