Reimplement overflow menu messages as DialogFragments
Bug: 20650813

This preserves fraction and "with leading digits" displays during
rotation.

It also turns out to easily support copy, which is a useful bonus,
since it was an obvious hole in the UI for the fraction display.
For the "with leading digits" display, this is similar to a
plain display copy, but it allows character-level selection.

Much of the code here was cloned from Justin's.

Change-Id: I4805280fa6a46f06833be0bde9563c3ce04dca45