Generalize % operator: Handle 100+10% as expected
Bug: 22836179

Handle + or - operators specially if they are followed by a constant
or pre-evaluated expression and the % operator.

Since this is basically an idiomatic use of the operators, we
apply it very conservatively.  When in doubt we use the old
interpretation.

Also fixes one unrelated anachronism in a comment.

Change-Id: I2f9fd26dd6c0456f0210815ef2bc1afca3a2b4d7
(cherry picked from commit 8afd0f85ed0b9fa1c96297c540cb74e6d8b9a64d)