Properly reserve display space for non-digits
Bug: 22882700

We did not correctly calculate the extra space we might need to add
a decimal point to a scientific notation results.  We also
incorrectly assumed that minus signs are the same width as a
digit. As a workaround we now reserve extra space for the possible
non-digit characters to ensure that a result is never inadvertently
truncated.

Change-Id: I79b499121703cd51bbbf3e1d85b2285a2dd5e025