Check for decimal point after adding ellipsis
Bug: 23728152

This avoids the one case in which we displayed neither a decimal
point nor an exponent, and the decimal point was not just to the
right of the number.

Remove a redundant test, which obscured the logic.

Add a couple of suggested manual tests that helped me find a
(newly introduced, not submitted) bug.

Change-Id: Iead1efa0a4e8ea74512e9e1f77334f80bbcdd202
(cherry picked from commit 73ecff20d9dfd80d1cff09e0210d8987049b9df3)