Update patch set 1
Patch Set 1: Code-Review-1

The black color is actually used to achieve a fade to black effect when you slide out the right panel containing the advanced functions.  Removing this causes the number pad to lighten rather than darken when this change is applied.

To keep the original behavior when no theme is applied, you should create a new color in XML and then load that color in code.  This way a theme can change that color if they choose to do so and it won't break the stock behavior.

Patch-set: 1