Handle font with missing FIGURE SPACE.
Bug: 27203752

Calculate digit width explicitly instead of relying on FIGURE SPACE.

This makes the calculator work correctly, even with fonts that are
utterly unreasonable for this purpose, like Nova Round. The result
remains ugly, due to highly variable digit widths. But it is correct.

This should have little to no impact on Nexus devices. Empirically
digit widths with our default font are slightly less than the width
of a FIGURE SPACE.  This version of the change does require careful
cross-device testing.

Change-Id: Id6feb0f312bc237ba462165ba2dff7030e04ae3b