Arrange for evaluateAndNotify calls at the right time.
Bug: 33011774
Bug: 33087975

We were sometimes calling it before layout after restart. Defer to
onLayout, as with requireResult.

Arrange for an explicit call when returning from HistoryFragment
to compensate for unintentional cancellation. This part is a hack
that needs to be revisited.

Adjust some constants to make it less likely that we will not see
am instant result, but equals will produce a valid result. This is
not a fundamental change, but 10000! now works as expected.
As does e^(1+10^-1000)-e, since we now force more evaluation to
try to distinguish a result from zero .  This may slow things down
on underpowered devices, but it shouldn't be a serious issue.

Change-Id: I5fcddef84dd907bc83d9bf575d0d378ca99d6359