Support accessibility in History.
Fixes: 33429660

Toggle accessibility importance of main Calculator elements based on
the open state of DragLayout. This is necessary due to the fact that
Talkback does not use visibility as a cue for determining accessibility
importance in RelativeLayout, which is the base class of DragLayout.
Without this workaround, it's possible to traverse to main Calculator
elements even when HistoryFragment is open.

Change-Id: Iff3d775ec72aa50fe8972c1def32f4999d90a8f9