Fix COPY string computation in exact case
Bug: 26175989
Bug: 33458621

Fix the determination of whether the currently displayed result is
already exact. The old logic failed to deal with the odd corner
case in which we force an initial scientific notation display because
we don't have room for digit separators.

When we recompute the exact result as part of a copy operation,
and the exact result is an integer, do not add a trailing decimal
point.

Change-Id: I3f673fc428cf6982eeaf6aa1037b9f522ea2d6f9