Correctly track Activity in Evaluator
Bug: 33703223

Evaluator was remembering the Activity with which it had been
created, not necessarily the current one. This interfered with
proper display of things like timeout messages that rely on the
remembered Activity.

Use the Application rather than Activity context for the DB helper,
since it outlives the Activity.

Remove unused mActivity from ExpressionDB.

Also acouple of trivial cleanups for minor annoyances found while
debugging:

1. Add some missing final declarations.

2. Fix a comment.

Change-Id: Iefe1fb588f66a1c77c82164680306377917c07af